Логика создания пользователя перенесена в сериализатор OrderWriteSerializer.
